src/main/java/dev/demo/spring5webfluxrest/services/impl/CategoryServiceImpl.java
package dev.demo.spring5webfluxrest.services.impl;

import dev.demo.spring5webfluxrest.commands.CategoryCommand;
import dev.demo.spring5webfluxrest.coverters.CategoryToCategoryCommandConverter;
import dev.demo.spring5webfluxrest.repositories.CategoryRepository;
import dev.demo.spring5webfluxrest.services.CategoryService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;

@Service
@RequiredArgsConstructor
public class CategoryServiceImpl implements CategoryService {

    private final CategoryRepository categoryRepository;
    private final CategoryToCategoryCommandConverter converter;

    @Override
    public Flux<CategoryCommand> findAll() {
        return categoryRepository.findAll().map(converter::convert);
    }

    @Override
    public Mono<CategoryCommand> findById(String id) {
        return categoryRepository.findById(id).map(converter::convert);
    }
}
